Face_Recognition_App/requirements.txt
numpy 
keras
keras_facenet
scikit_learn
pandas
opencv-python
mtcnn
tensorflow
streamlit
pillow
-e .